ENH: move gamma responsibility to each backend
